services/backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = "Article Information Parser backend"
authors = ["Gilles Magalhaes <gilles.magalhaes@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.1.2"
djangorestframework = "^3.14.0"
drf-spectacular = "^0.24.2"
django-filter = "^22.1"
networkx = "^2.8.7"
network = "^0.1"
numpy = "^1.23.4"
scipy = "^1.9.3"
igraph = "^0.10.2"
#leidenalg = "^0.9.0"
#sklearn = "^0.0"
django-cors-headers = "^3.13.0"
psycopg2-binary = "^2.9.4"
matplotlib = "^3.6.1"
requests = "^2.28.1"
beautifulsoup4 = "^4.11.1"
django-environ = "^0.9.0"
gunicorn = "^20.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
